pyproject.toml
[project]
name = "ai-toolbox"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "jupyterlab",
    "python-dotenv",
    "streamlit",
    "openai>=2.31.0",
    "gradio>=6.11.0",
    "ipywidgets>=8.1.8",
    "mlx-whisper>=0.4.3",
    "openai-agents>=0.14.4",
    "sendgrid>=6.12.5",
    "anthropic>=0.49.0",
]
